requirements.txt
# Requirements for the sylistock Django project
# Target Python 3.12+. Pin or adjust versions for production as needed.

Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9
gunicorn>=20.1; platform_system != "Windows"
whitenoise>=6.0
django-cors-headers>=4.0

# Testing
pytest>=8.3
pytest-django>=4.0
pytest-cov>=3.0
coverage>=6.0

# End of file
